# src/main/resources/api.yaml
openapi: "3.0.3"
info:
  title: Biblioteca API
  description: API para gerenciar organizações
  version: 1.0.0
...
paths:
...
  /organizacao/find:
    post:
      summary: Busca Organizações
      security:
        - bearerAuth: []
      tags:
        - Organização
      operationId: findOrganizacao
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizacaoRequest'
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizacaoResponseDTO'
  /organizacao/detail/{id}:
    get:
      summary: Detalha  uma organização
      security:
        - bearerAuth: []
      tags:
        - Organização
      operationId: detailOrganizacao
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizacaoResponseDTO'
  /organizacao:
    put:
      summary: Cria uma organização
      security:
        - bearerAuth: []
      tags:
        - Organização
      operationId: createOrganizacao
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizacaoRequestDTO'
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizacaoModel'
    patch:
      summary: Atualiza uma organização
      security:
        - bearerAuth: []
      tags:
        - Organização
      operationId: updateOrganizacao
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizacaoModel'
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizacaoModel'
  /organizacao/delete/{id}:
    delete:
      summary: Deleta uma organização
      security:
        - bearerAuth: []
      tags:
        - Organização
      operationId: deleteOrganizacao
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content
...
  /endereco/{id}:
    put:
      summary: Atualiza um endereço existente
      operationId: updateEndereco
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnderecoModel'
      responses:
        '200':
          description: Endereço atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnderecoModel'
        '404':
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Endereço não encontrado."
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Erro interno."
    get:
      summary: Detalha um endereço
      security:
        - bearerAuth: []
      tags:
        - Endereço
      operationId: detailEndereco
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnderecoModel'
    delete:
      summary: Deleta um endereço existente
      operationId: deleteEndereco
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Endereço excluído com sucesso
        '404':
          description: Endereço não encontrado
        '500':
          description: Erro interno do servidor
...
  /endereco2:
    post:
      summary: Cria um novo endereço
      security:
        - bearerAuth: []
      tags:
        - Endereço
      operationId: createEndereco2
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Endereco2Model'
      responses:
        '200':
          description: Endereço criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Endereco2Model'
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Endereço não criado."
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Erro interno."
    get:
      summary: Busca Enderecos
      security:
        - bearerAuth: []
      tags:
        - Perfil
      operationId: findEndereco2
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Endereco2Model'
  /endereco2/{id}:
    put:
      summary: Atualiza um endereço existente
      operationId: updateEndereco2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Endereco2Model'
      responses:
        '200':
          description: Endereço atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Endereco2Model'
        '404':
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Endereço não encontrado."
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Erro interno."
    get:
      summary: Detalha um endereço
      security:
        - bearerAuth: []
      tags:
        - Endereço
      operationId: detailEndereco2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Endereco2Model'
    delete:
      summary: Deleta um endereço existente
      operationId: deleteEndereco2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Endereço excluído com sucesso
        '404':
          description: Endereço não encontrado
        '500':
          description: Erro interno do servidor
  /pais:
    get:
      summary: Busca Países
      security:
        - bearerAuth: []
      tags:
        - País
      operationId: findPais
      parameters:
        - name: nome
          in: path
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaisModel'
  /pais/{id}:
    get:
      summary: Detalha um País
      security:
        - bearerAuth: []
      tags:
        - País
      operationId: detailPais
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaisModel'
  /estado:
    get:
      summary: Busca Estados
      security:
        - bearerAuth: []
      tags:
        - Estado
      operationId: findEstado
      parameters:
        - name: nome
          in: path
          required: false
          schema:
            type: string
        - name: pais_id
          in: path
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EstadoModel'
  /estado/{id}:
    get:
      summary: Detalha um Estado
      security:
        - bearerAuth: []
      tags:
        - Estado
      operationId: detailEstado
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstadoModel'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT  # Especifica o formato do token, pode ser omitido
  schemas:
...
    OrganizacaoRequest:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        nif:
          type: string
        email:
          type: string
        website:
          type: string
        setorAtividade:
          type: string
        limit:
          type: integer
        offset:
          type: integer
        event:
          type: object
          $ref: '#/components/schemas/EventPagination'
    OrganizacaoRequestDTO:
        type: object
        properties:
            id:
              type: integer
              description: Identificador da organização
              example: 1
            nome:
              type: string
              description: Nome da organização
              example: "XPTO Lda"
            nif:
              type: string
              description: Número de Identificação Fiscal
              example: "123456789"
            email:
              type: string
              description: Endereço de e-mail
              example: "Baker Street 221B"
            website:
              type: string
              description: Endereço do website
              example: "https://www.exemplo.com"
            setorAtividade:
              type: string
              description: Setor de atividade
              example: "Educação"
            missao:
              type: string
              description: Missão da organização
              example: "Promover a educação"
            representanteLegal:
              type: string
              description: Representante legal
              example: "João Silva"
            cargo:
              type: string
              description: Cargo do representante legal
              example: "CEO"
            numeroRegistoComercial:
              type: string
              description: Número de registo comercial
              example: "123456789"
            dataRegisto:
              type: string
              format: date
              nullable: true
              description: Data de registo comercial
              example: "2021-09-01"
...
    EventPagination:
      type: object
      properties:
        length:
          type: integer
        pageIndex:
          type: integer
        pageSize:
            type: integer
        previousPageIndex:
            type: integer
...
    OrganizacaoResponseDTO:
      type: object
      properties:
        organizacoes:
          type: array
          items:
            $ref: '#/components/schemas/OrganizacaoModel'
        totalRecords:
          type: integer
...
    OrganizacaoModel:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        nif:
          type: string
        email:
          type: string
        website:
          type: string
        setorAtividade:
            type: string
        missao:
            type: string
        representante_legal:
            type: string
        cargo:
            type: string
        numeroRegistoComercial:
            type: string
        dataRegisto:
            type: string
            format: date
            nullable: true
...
    EnderecoModel:
      type: object
      properties:
        id:
          type: integer
        logradouro:
          type: string
        numero:
          type: integer
        complemento:
            type: string
        bairro:
          type: string
        cidade:
          type: string
        uf:
          $ref: '#/components/schemas/UnidadeFederativaModel'
        cep:
          type: string
    Endereco2Model:
      type: object
      properties:
        id:
          type: integer
        cidade_id:
          $ref: '#/components/schemas/CidadeModel'
        cep:
          type: string
        logradouro:
          type: string
        numero:
          type: string
        complemento:
          type: string
        bairro:
          type: string
        latitude:
          type: number
        longitude:
            type: number
        usuario_id:
          type: integer
...
    UnidadeFederativaModel:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        sigla:
            type: string
    CidadeModel:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        estado_id:
          $ref: '#/components/schemas/EstadoModel'
    EstadoModel:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        pais_id:
          $ref: '#/components/schemas/PaisModel'
    PaisModel:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string